日志中心监控脚本/requirements.txt
Flask==2.0.1
PyYAML==6.0.1
requests==2.28.1
websocket_client==1.6.2
